@@ -1,10 +1,10 @@
 name: titled_navigation_bar
 description: A beautiful, clean and simple bottom navigation bar with smooth animation on click. This package is high customizable, read more bellow for more details.
-version: 4.1.0
+version: 5.0.0-nullsafety.0
 homepage: https://github.com/pedromassango/titled_navigation_bar
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
